Added createDatabaseIfNotExists(..) methods to CouchDbInstance.
Returns true if database was created and false if database already exists.

Using this method in createDatabase(DbPath) to get rid of the checkIfDbExists call. This prevents a race condition that could happen if the database was created by someone else between the checkIfDbExists call and the restTemplate.put call.