Cross-Platform New Line for SimpleViewGenerator
Running a maven build on non-*nix machines fails SimpleViewGeneratorTest
because of line endings. This functionality already works on Windows,
but tests fail because of a reference to `String#replaceAll("\n", "")`.
This changes defines the cross-platform `String.format("%n")` as a
constant, and references it in the `replaceAll` call as per
http://docs.oracle.com/javase/7/docs/api/java/util/Formatter.html#syntax